# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Billing & Solar ERP API",
        "description": "Billing + Cloud Solar Panel Installation & Service Management API. Workflow: Lead → Survey → Quotation → Installation → Service & AMC. Roles: MD Super Admin, Branch Manager, Branch Admin, Sales Executive, Site Survey Engineer, Installation Manager/Technician, Service Team, Accounts. JWT Secured: use POST /api/v1/auth/login then header Authorization: Bearer {access_token}. AI-friendly: structured JSON responses, clear field names, dashboard metrics.",
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server (localhost)"
        },
        {
            "url": "https://testbillapi.eazycutz.com/api/v1",
            "description": "API Server (Live)"
        },
        {
            "url": "/api/v1",
            "description": "API Server (relative)"
        }
    ],
    "paths": {
...
        "/super-admin/dashboard": {
            "get": {
                "tags": [
                    "Super Admin"
                ],
                "summary": "Super Admin Dashboard",
                "description": "Super Admin Dashboard - SaaS-wide stats.",
                "operationId": "7983f0b6f94430d4fad8d13eda3cf6fa",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "month",
                            "enum": [
                                "today",
                                "week",
                                "month",
                                "year"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_organizations": {
                                            "type": "integer"
                                        },
                                        "active_organizations": {
                                            "type": "integer"
                                        },
                                        "total_users": {
                                            "type": "integer"
                                        },
                                        "total_sales": {
                                            "type": "number"
                                        },
                                        "paid": {
                                            "type": "number"
                                        },
                                        "unpaid": {
                                            "type": "number"
                                        },
                                        "period": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/super-admin/login-as/{user}": {
            "post": {
                "tags": [
                    "Super Admin"
                ],
                "summary": "Login As User (Impersonate)",
                "description": "Super admin obtains a JWT token for the target user (by user ID) to access their organization. Use the returned access_token as Bearer for subsequent requests as that user.",
                "operationId": "d196d95edcd19c2f4dde0f790eb94a43",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Target user ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        },
                                        "impersonating": {
                                            "description": "Target user info",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: \\\"Authorization: Bearer {token}\\\"",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}